refactor index.ks to display error msg
