+++ src/view/error.js
@@ -1,5 +1,4 @@
-const error = 
-`
+const error = `
   <div class="alert alert-danger" id="displayError" role="alert" style="display: none"></div>
 `;
